requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
python-multipart==0.0.9
anthropic==0.18.1
streamlit==1.32.0
langchain==0.1.12
langchain-openai==0.0.8
langchain-google-genai==0.0.11
langchain-anthropic==0.1.4
langchain-community>=0.0.28
chromadb==0.4.24
python-dotenv==1.0.1
google-generativeai==0.3.2
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.5.1
python-docx==1.1.0
PyPDF2==3.0.1
SQLAlchemy==2.0.28
pymysql==1.1.0
feedparser==6.0.10
beautifulsoup4==4.12.3
aiohttp==3.9.3